+++ src/toColNums.cpp
@@ -47,7 +47,6 @@ std::vector<double> toColNums(Eigen::SparseMatrix<double> data) {
 std::vector<double> ColNotZero(Eigen::SparseMatrix<double> data) {
   std::vector<double> tmp( data.outerSize() );
   int iter=0;
-  int id = 0;
   for ( int k=0; k < data.outerSize(); ++k){
     tmp[iter] = 0;
     for (Eigen::SparseMatrix<double>::InnerIterator it(data, k); it; ++it){